+++ lib/marshalling.dart
@@ -3,6 +3,7 @@ import 'dart:collection';
 part 'src/marshalling/iterable_type_info.dart';
 part 'src/marshalling/map_type_info.dart';
 part 'src/marshalling/marshaller.dart';
+part 'src/marshalling/marshalling_error.dart';
 part 'src/marshalling/property_accessor.dart';
 part 'src/marshalling/property_info.dart';
 part 'src/marshalling/type_info.dart';