+++ lib/src/marshalling/marshalling_error.dart
@@ -1,0 +1,16 @@
+part of '../../marshalling.dart';
+
+class MarshallingError {
+  final String message;
+
+  MarshallingError(this.message);
+
+  @override
+  String toString() {
+    if (message == null) {
+      return 'Marshalling error';
+    }
+
+    return 'Marshalling error: $message';
+  }
+}